requirements.txt
flask
requests
fpdf
python-dotenv
flasgger
gunicorn
anthropic

